refactor: fix circular dependency
The command that shows the login vue was having circular
dependency issues.

This moves that logic in to its own file to fix this.
