+++ packages/core/src/auth/activation.ts
@@ -11,8 +11,8 @@ import { getLogger } from '../shared/logger'
 import { ExtensionUse, initAuthCommands } from './utils'
 import { isCloud9 } from '../shared/extensionUtilities'
 import { isInDevEnv } from '../shared/vscode/env'
-import { registerCommands, getShowManageConnections } from './ui/vue/show'
 import { isWeb } from '../shared/extensionGlobals'
+import { getShowManageConnections, registerCommands } from '../login/command'
 
 export async function initialize(
     extensionContext: vscode.ExtensionContext,
